Move query() method from api.py to query.py
